requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
autopep8==2.3.1
bcrypt==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
confluent-kafka==2.6.0
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-utils==0.8.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
platformdirs==4.3.6
psutil==5.9.8
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomlkit==0.13.2
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
